itsquizz-be/requirements.txt
flask 
# flask_sqlalchemy 
flask_session 
werkzeug
python-dotenv
flask-session 
sqlalchemy
flask-cors